backend/9/Mnasa/Newmnasa/requirements.txt
Django==4.2.20
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
Pillow==10.2.0
python-dotenv==1.0.1
djangorestframework-simplejwt==5.3.0
boto3==1.34.34
botocore==1.34.34
ffmpeg-python==0.2.0
python-magic==0.4.27 